crypto: new base interface
This introduces the simplified CryptoBase interface that will replace
the CryptoImpl interface
